Initial Commit
The kalmpr.c program computes and prints the number of all k-almost prime numbers found within the interval [MINNUM,MAXNUM], for 2≤ k ≤ MAXK, and the number of those that are in particular, 2-relative prime numbers, i.e., half-primes. For the half-integers of interest, it prints the factors of which they are decomposed. Then we increase the range of the range of the [MINNUM,MAXNUM] interval to see the rate at which the execution time of our program also increases. Then the program selects 20 positive integer numbers for each of which it prints the k corresponding to it.